Number of enhancements including: - Storage capacity indicator (progress bar) in storage menu with text count - Storage capacity indicator on circle menu button - UI improvements to support various resolutions (rather than just my android) - CD Model (first time using blender) - Tweaking lighting

+++ Assets/Scripts/Models/Device.cs
@@ -65,7 +65,7 @@ public class Device : Purchaseable {
 	}
 
 	public string GetDescription() {
-		return "Outbound BPS: " + BitUtil.StringFormat(OutBps, BitUtil.TextFormat.Long, true);
+		return "Outbound BPS: " + BitUtil.StringFormat(OutBps, BitUtil.TextFormat.Short, true);
 	}
 
 	public float GetCost() {
